test: add autocomplete test
