+++ extensions/intellij/.gitignore
@@ -4,3 +4,5 @@ src/main/resources/bin
 src/main/resources/config_schema.json
 src/main/resources/continue_rc_schema.json
 video
+src/test/kotlin/com/github/continuedev/continueintellijextension/e2e/test-continue/*
+!src/test/kotlin/com/github/continuedev/continueintellijextension/e2e/test-continue/config.json
