--- extensions/intellij/.run/Run IDE for UI Tests.run.xml
@@ -1,25 +1,0 @@
-<!--<component name="ProjectRunConfigurationManager">-->
-<!--  <configuration default="false" name="Run IDE for UI Tests" type="GradleRunConfiguration" factoryName="Gradle">-->
-<!--    <log_file alias="idea.log" path="$PROJECT_DIR$/build/idea-sandbox/system/log/idea.log" />-->
-<!--    <ExternalSystemSettings>-->
-<!--      <option name="executionName" />-->
-<!--      <option name="externalProjectPath" value="$PROJECT_DIR$" />-->
-<!--      <option name="externalSystemIdString" value="GRADLE" />-->
-<!--      <option name="scriptParameters" value="" />-->
-<!--      <option name="taskDescriptions">-->
-<!--        <list />-->
-<!--      </option>-->
-<!--      <option name="taskNames">-->
-<!--        <list>-->
-<!--          <option value="runIdeForUiTests" />-->
-<!--        </list>-->
-<!--      </option>-->
-<!--      <option name="vmOptions" />-->
-<!--    </ExternalSystemSettings>-->
-<!--    <ExternalSystemDebugServerProcess>true</ExternalSystemDebugServerProcess>-->
-<!--    <ExternalSystemReattachDebugProcess>true</ExternalSystemReattachDebugProcess>-->
-<!--    <DebugAllEnabled>false</DebugAllEnabled>-->
-<!--    <RunAsTest>false</RunAsTest>-->
-<!--    <method v="2" />-->
-<!--  </configuration>-->
-<!--</component>-->
