+++ extensions/intellij/src/test/kotlin/com/github/continuedev/continueintellijextension/e2e/Autocomplete.kt
@@ -1,0 +1,73 @@
+package com.github.continuedev.continueintellijextension.e2e
+
+import com.automation.remarks.junit5.Video
+import com.github.continuedev.continueintellijextension.fixtures.dialog
+import com.github.continuedev.continueintellijextension.fixtures.idea
+import com.github.continuedev.continueintellijextension.fixtures.welcomeFrame
+import com.github.continuedev.continueintellijextension.utils.RemoteRobotExtension
+import com.github.continuedev.continueintellijextension.utils.StepsLogger
+import com.intellij.remoterobot.RemoteRobot
+import com.intellij.remoterobot.fixtures.ComponentFixture
+import com.intellij.remoterobot.search.locators.byXpath
+import com.intellij.remoterobot.steps.CommonSteps
+import com.intellij.remoterobot.utils.keyboard
+import com.intellij.remoterobot.utils.waitFor
+import com.intellij.remoterobot.utils.waitForIgnoringError
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
+import java.time.Duration.ofMinutes
+import java.time.Duration.ofSeconds
+
+@ExtendWith(RemoteRobotExtension::class)
+class Autocomplete {init {
+    StepsLogger.init()
+}
+
+    @BeforeEach
+    fun waitForIde(remoteRobot: RemoteRobot) {
+        waitForIgnoringError(ofMinutes(3)) { remoteRobot.callJs("true") }
+    }
+
+    @AfterEach
+    fun closeProject(remoteRobot: RemoteRobot) = CommonSteps(remoteRobot).closeProject()
+
+    @Test
+    @Video
+    fun displayCompletion(remoteRobot: RemoteRobot): Unit = with(remoteRobot) {
+        welcomeFrame {
+            createNewProjectLink.click()
+            dialog("New Project") {
+                findText("Java").click()
+                checkBox("Add sample code").select()
+                button("Create").click()
+            }
+        }
+
+        // Wait for the default "Main.java" tab to load
+        // Our "continue_tutorial.java.ft" tab loads first, but then "Main.java" takes focus.
+        // So we need to wait for that to occur, and then focus on "continue_tutorial.java.ft"
+        waitFor(ofSeconds(20)) {
+            findAll<ComponentFixture>(
+                byXpath("//div[@accessiblename='Main.java' and @class='EditorTabLabel']")
+            ).isNotEmpty()
+        }
+
+        idea {
+            with(textEditor()) {
+                val USER_MESSAGE = "TEST_USER_MESSAGE_0"
+                editor.insertTextAtLine(0, 0, USER_MESSAGE)
+                editor.clickOnOffset(USER_MESSAGE.length)
+
+                keyboard {
+                    enterText(" ")
+                }
+
+                waitFor(ofSeconds(10)) {
+                    editor.hasText("TEST_LLM_RESPONSE_0")
+                }
+            }
+        }
+    }
+}
