+++ extensions/intellij/src/test/kotlin/com/github/continuedev/continueintellijextension/e2e/test-continue/config.json
@@ -1,0 +1,43 @@
+{
+  "ui": {
+    "displayRawMarkdown": false
+  },
+  "models": [
+    {
+      "title": "TEST LLM",
+      "provider": "test",
+      "model": "this field is not used"
+    },
+    {
+      "title": "Mock",
+      "provider": "mock",
+      "model": "this field is not used"
+    }
+  ],
+  "analytics": {
+    "provider": "continue-proxy"
+  },
+  "tabAutocompleteModel": {
+    "title": "TEST LLM",
+    "provider": "test",
+    "model": "this field is not used"
+  },
+  "tabAutocompleteOptions": {
+    "useCache": false
+  },
+  "contextProviders": [
+    {
+      "name": "docs"
+    },
+    {
+      "name": "diff"
+    },
+    {
+      "name": "url"
+    },
+    {
+      "name": "folder"
+    }
+  ],
+  "docs": []
+}
