+++ gui/src/components/dialogs/ConfirmationDialog.tsx
@@ -17,7 +17,7 @@ function ConfirmationDialog(props: ConfirmationDialogProps) {
 
   return (
     <div className="p-4 pt-0">
-      <h1 className="text-2x mb-1 text-center">
+      <h1 className="mb-1 text-center text-xl">
         {props.title ?? "Confirmation"}
       </h1>
       <p className="text-center text-base" style={{ whiteSpace: "pre-wrap" }}>
